[traserver] First idea of what a traverser can be
The idea is to keep the DOM tree has separate as possible from the
clutter scene graph. To build the scene graph a "Traverser" is
instanciated and used to walk through the DOM tree, building actors
along the way.

With all this, we can draw a rectangle. How many lines of code do we
need to draw rectangles these days? A lot.