[parser] Fix parsing of unknown and empty elements
when we encountered an unknown element, we had the start/end balance
wrong. Empty elements were an issue too an the xml reader don't emit a
END_ELEMENT on those nodes.

In the process, it was interesting to discover that even the tree
walking was wrong.

Oh, and it's a good occasion to put back the non-hacky svg tests from
the tiny 1.2 spec.

So basically, everything was wrong but worked on small test cases.