parser: handle the parsing of namespace declarations
We are now able to do something with namespaces declarations: we keep
the tuple (uri, prefix) inside DomDocument, every property should have
a namespaceURI associated with it, Element::setAttributeNS() is now
abstracted in DomElement and subclasses do not have to do anything to
support that.

node::ownerDocument is now implemented for element et text nodes.