udom: Introduce SVTimer
First standalone object of the udom. A big issue remains:
addEventListener() takes a function in javascript, which is, well, not
handled at all by the XmlEventTarget interface (which takes an object
implementing the XmlEventListener interface). Gjs bails out as it's not
able to convert a JS Function to a XmlEventListener.