tests: Make Javascript unit tests possible
This is done by importing jsUnitCore.js from the jsUnit project and
writing a small wrapper to launch the tests.

Along the way, libdox-0.2 is born, it's the low lovel DOM library that
defines the navigator object (originally needed to evaluate
jsUnitCore.js). A following commit will migrate DaxDom* to Dox*.

The JS context is now tied to the document as it should have been since
the start and not a global singleton. test-js needed this to access the
JS context once the SVG document is parsed to evaluate daxSucess and
retrieve the success state of the tests.