traverser: Cleanup and new traverser ("load" handlers)
test-js needs to handle <handler> elements with the "load" event to fire
the main test function. Let's define a new traverser called
traverser-load whose only job is to find <handler> elements with
type="load" and run them.

While at it, clean-up a bit the traverser class: It's an abstract class
now, there's no reason to be able to instanciate it. "root" is a new
property on that class to be able to define _with_root() constructors in
a more gobjecty fashion (no code, just a g_object_new() call).